requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
streamlit>=1.30.0
pydantic>=2.0.0
pyyaml>=6.0
dashscope>=1.14.0
